Extract AccountTitle from AccountHeaderCard to prevent re-rendering MainTitle
