+++ src/components/Account/AccountHeaderCard.tsx
@@ -6,24 +6,19 @@ import CardContent from "@material-ui/core/CardContent"
 import Dialog from "@material-ui/core/Dialog"
 import IconButton from "@material-ui/core/IconButton"
 import Slide from "@material-ui/core/Slide"
-import Tooltip from "@material-ui/core/Tooltip"
-import GroupIcon from "@material-ui/icons/Group"
 import MoreVertIcon from "@material-ui/icons/MoreVert"
 import SwapHorizIcon from "@material-ui/icons/SwapHoriz"
-import VerifiedUserIcon from "@material-ui/icons/VerifiedUser"
 import { Account, AccountsContext, AccountsContextType } from "../../context/accounts"
 import { SettingsContext } from "../../context/settings"
 import { useAccountData, useIsMobile, useRouter } from "../../hooks"
 import * as routes from "../../routes"
-import { primaryBackgroundColor } from "../../theme"
 import AccountDeletionDialog from "../Dialog/AccountDeletion"
 import ChangePasswordDialog from "../Dialog/ChangePassword"
 import ExportKeyDialog from "../Dialog/ExportKey"
 import RenameDialog from "../Dialog/Rename"
 import ButtonIconLabel from "../ButtonIconLabel"
-import { HorizontalLayout } from "../Layout/Box"
-import MainTitle from "../MainTitle"
 import AccountContextMenu from "./AccountContextMenu"
+import AccountTitle from "./AccountTitle"
 
 enum DialogID {
   changePassword,
@@ -33,31 +28,6 @@ enum DialogID {
 }
 
 const DialogTransition = (props: any) => <Slide {...props} direction="up" />
-
-function PasswordStatus(props: { safe: boolean; style?: React.CSSProperties }) {
-  return (
-    <Tooltip title={props.safe ? "Password protected" : "No password"}>
-      <VerifiedUserIcon style={{ opacity: props.safe ? 1 : 0.5, ...props.style }} />
-    </Tooltip>
-  )
-}
-
-function TestnetBadge(props: { style?: React.CSSProperties }) {
-  const style: React.CSSProperties = {
-    display: "inline-flex",
-    alignItems: "center",
-    padding: "5px",
-    background: "white",
-    borderRadius: 3,
-    color: primaryBackgroundColor,
-    fontSize: "50%",
-    fontWeight: "bold",
-    lineHeight: "100%",
-    textTransform: "uppercase",
-    ...props.style
-  }
-  return <span style={style}>Testnet</span>
-}
 
 interface Props {
   account: Account
@@ -71,8 +41,8 @@ interface Props {
 function AccountHeaderCard(props: Props) {
   const { changePassword, removePassword } = React.useContext(AccountsContext)
   const isSmallScreen = useIsMobile()
-  const settings = React.useContext(SettingsContext)
   const router = useRouter()
+  const settings = React.useContext(SettingsContext)
 
   const [openDialog, setOpenDialog] = React.useState<DialogID | null>(null)
   const accountData = useAccountData(props.account.publicKey, props.account.testnet)
@@ -82,6 +52,47 @@ function AccountHeaderCard(props: Props) {
   const performRenaming = React.useCallback((newName: string) => props.onRenameAccount(props.account.id, newName), [
     props.account.id
   ])
+
+  const actions = React.useMemo(
+    () => (
+      <>
+        <Button
+          onClick={() => router.history.push(routes.tradeAsset(props.account.id))}
+          style={{
+            borderColor: "rgba(255, 255, 255, 0.9)",
+            color: "white",
+            padding: "0 12px",
+            marginRight: isSmallScreen ? 0 : 8,
+            minHeight: 36
+          }}
+          variant="outlined"
+        >
+          <ButtonIconLabel label="Trade">
+            <SwapHorizIcon />
+          </ButtonIconLabel>
+        </Button>
+
+        <AccountContextMenu
+          account={props.account}
+          activated={accountData.activated}
+          settings={settings}
+          onChangePassword={() => setOpenDialog(DialogID.changePassword)}
+          onDelete={() => setOpenDialog(DialogID.deleteAccount)}
+          onExport={() => setOpenDialog(DialogID.exportKey)}
+          onManageAssets={props.onManageAssets}
+          onManageSigners={props.onManageSigners}
+          onRename={() => setOpenDialog(DialogID.renameAccount)}
+        >
+          {({ onOpen }) => (
+            <IconButton color="inherit" onClick={onOpen} style={{ marginRight: -16, fontSize: 32 }}>
+              <MoreVertIcon style={{ fontSize: "inherit" }} />
+            </IconButton>
+          )}
+        </AccountContextMenu>
+      </>
+    ),
+    [props.account, accountData.activated, settings, props.onManageAssets, props.onManageSigners]
+  )
 
   return (
     <Card
@@ -94,60 +105,7 @@ function AccountHeaderCard(props: Props) {
       }}
     >
       <CardContent style={isSmallScreen ? { padding: 8 } : undefined}>
-        <MainTitle
-          title={<span style={{ marginRight: 20 }}>{props.account.name}</span>}
-          titleColor="inherit"
-          onBack={() => router.history.push(routes.allAccounts())}
-          style={{ marginTop: -12, marginLeft: 0 }}
-          badges={
-            <HorizontalLayout display="inline-flex" alignItems="center" width="auto" fontSize="1.5rem">
-              {props.account.testnet ? <TestnetBadge style={{ marginRight: 16 }} /> : null}
-              {accountData.signers.length > 1 ? (
-                <Tooltip title="Multi-Signature Account">
-                  <GroupIcon style={{ fontSize: "120%", marginRight: 8 }} />
-                </Tooltip>
-              ) : null}
-              <PasswordStatus safe={props.account.requiresPassword} style={{ fontSize: "90%", marginTop: "-0.05em" }} />
-            </HorizontalLayout>
-          }
-          actions={
-            <>
-              <Button
-                onClick={() => router.history.push(routes.tradeAsset(props.account.id))}
-                style={{
-                  borderColor: "rgba(255, 255, 255, 0.9)",
-                  color: "white",
-                  padding: "0 12px",
-                  marginRight: isSmallScreen ? 0 : 8,
-                  minHeight: 36
-                }}
-                variant="outlined"
-              >
-                <ButtonIconLabel label="Trade">
-                  <SwapHorizIcon />
-                </ButtonIconLabel>
-              </Button>
-
-              <AccountContextMenu
-                account={props.account}
-                activated={accountData.activated}
-                settings={settings}
-                onChangePassword={() => setOpenDialog(DialogID.changePassword)}
-                onDelete={() => setOpenDialog(DialogID.deleteAccount)}
-                onExport={() => setOpenDialog(DialogID.exportKey)}
-                onManageAssets={props.onManageAssets}
-                onManageSigners={props.onManageSigners}
-                onRename={() => setOpenDialog(DialogID.renameAccount)}
-              >
-                {({ onOpen }) => (
-                  <IconButton color="inherit" onClick={onOpen} style={{ marginRight: -16, fontSize: 32 }}>
-                    <MoreVertIcon style={{ fontSize: "inherit" }} />
-                  </IconButton>
-                )}
-              </AccountContextMenu>
-            </>
-          }
-        />
+        <AccountTitle account={props.account} accountData={accountData} actions={actions} />
 
         {props.children}
